Backend/category_service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==3.0.1
requests==2.31.0
gunicorn==20.1.0
